refactor: rename the specs command group module data.py → specs.py
The `specs` command group lived in `data.py` (a legacy name from when the
command was `data`). Rename it to `specs.py` for consistency with the
command name.

Since `specs.py` already existed (the `validate` command + flat-layout
helpers that the group imports), it's renamed to `specs_layout.py` — kept
callback-free so it stays unit-testable in isolation. The group (now
`specs.py`) imports `validate` from it; `cli.py` wires `specs.app`;
`test_specs.py` targets `specs_layout`.

No behavior change — pure module relocation + import-path updates.
